requirements.txt
python-dotenv
csv
smtplib
EmailMessage
mimetypes